Refactor TourDetailsPage and ToursPage to use context for trek data
- Integrated `useTours` context in `TourDetailsPage` to fetch trek details by ID.
- Removed hardcoded trek data from `TourDetailsPage` and `ToursPage`, utilizing context for filtered treks.
- Updated `Trek` type to include `slideshowImages` for better data structure.
- Simplified state management by eliminating unnecessary local state in `ToursPage`.
- Adjusted carousel image handling in `TourDetailsPage` to use `slideshowImages` from context.
