+++ frontend/src/App.tsx
@@ -2,17 +2,20 @@ import { HashRouter as Router, Routes, Route } from 'react-router-dom';
 import HomePage from './pages/HomePage';
 import ToursPage from './pages/ToursPage';
 import TourDetailsPage from './pages/TourDetailsPage';
+import { TourProvider } from './contexts/TourContext';
 import './styles/custom.css';
 
 function App() {
   return (
-    <Router>
-      <Routes>
-        <Route path="/" element={<HomePage />} />
-        <Route path="/tours" element={<ToursPage />} />
-        <Route path="/tours/:id" element={<TourDetailsPage />} />
-      </Routes>
-    </Router>
+    <TourProvider>
+      <Router>
+        <Routes>
+          <Route path="/" element={<HomePage />} />
+          <Route path="/tours" element={<ToursPage />} />
+          <Route path="/tours/:id" element={<TourDetailsPage />} />
+        </Routes>
+      </Router>
+    </TourProvider>
   );
 }
